Rename package company.
