+++ app/src/main/java/com/guagua/githuber_android/domain/Extension.kt
@@ -1,0 +1,7 @@
+package com.guagua.githuber_android.domain
+
+import com.guagua.githuber_android.model.data.User
+
+fun User.toUserItem(): UserItem {
+    return UserItem(id = id, name = login, avatar = avatarUrl)
+}
