+++ app/src/main/java/com/guagua/githuber_android/domain/UserItem.kt
@@ -1,0 +1,3 @@
+package com.guagua.githuber_android.domain
+
+data class UserItem(val id: Int, val name: String, val avatar: String)
