+++ app/src/main/java/com/guagua/githuber_android/model/data/Collection.kt
@@ -1,0 +1,8 @@
+package com.guagua.githuber_android.model.data
+
+import com.google.gson.annotations.SerializedName
+
+data class Collection<T>(
+    @SerializedName("total_count") val totalCount: Int,
+    @SerializedName("items") val items: List<T>
+)
