+++ app/src/main/java/com/guagua/githuber_android/model/data/User.kt
@@ -1,0 +1,12 @@
+package com.guagua.githuber_android.model.data
+
+import com.google.gson.annotations.SerializedName
+
+data class User(
+    @SerializedName("login") val login: String,
+    @SerializedName("id") val id: Int,
+    @SerializedName("avatar_url") val avatarUrl: String,
+    @SerializedName("url") val url: String,
+    @SerializedName("type") val type: String,
+    @SerializedName("score") val score: Float
+)
