+++ app/src/main/java/com/guagua/githuber_android/model/user/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.guagua.githuber_android.model.user
+
+
+class UserRepository(private val dataSource: UserDataSource) {
+    val apiResponse = dataSource.apiResponse
+
+    fun searchUsers(query: String, page: Int, perPage: Int) {
+        dataSource.searchUsers(query, page, perPage)
+    }
+}
