+++ app/src/main/java/com/guagua/githuber_android/ui/ChangeType.kt
@@ -1,0 +1,5 @@
+package com.guagua.githuber_android.ui
+
+enum class ChangeType {
+    ADDED, MODIFIED
+}
