+++ app/src/main/java/com/guagua/githuber_android/ui/SearchListViewModelFactory.kt
@@ -1,0 +1,14 @@
+package com.guagua.githuber_android.ui
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.guagua.githuber_android.model.user.UserRepository
+
+class SearchListViewModelFactory(
+    private val repository: UserRepository
+) : ViewModelProvider.NewInstanceFactory() {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        return SearchListViewModel(repository) as T
+    }
+}
