--- app/src/main/java/com/kkbox/githuber_android/domain/UserItem.kt
@@ -1,3 +1,0 @@
-package com.kkbox.githuber_android.domain
-
-data class UserItem(val id: Int, val name: String, val avatar: String)
