--- app/src/main/java/com/kkbox/githuber_android/model/user/UserDataSource.kt
@@ -1,10 +1,0 @@
-package com.kkbox.githuber_android.model.user
-
-import com.kkbox.githuber_android.model.common.ApiResponse
-import com.kkbox.githuber_android.model.common.SingleLiveEvent
-
-interface UserDataSource {
-    val apiResponse: SingleLiveEvent<ApiResponse>
-
-    fun searchUsers(query: String, page: Int, perPage: Int)
-}
