--- app/src/main/java/com/kkbox/githuber_android/ui/ChangeType.kt
@@ -1,5 +1,0 @@
-package com.kkbox.githuber_android.ui
-
-enum class ChangeType {
-    ADDED, MODIFIED
-}
