--- app/src/main/java/com/kkbox/githuber_android/ui/SearchListViewModelFactory.kt
@@ -1,14 +1,0 @@
-package com.kkbox.githuber_android.ui
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.kkbox.githuber_android.model.user.UserRepository
-
-class SearchListViewModelFactory(
-    private val repository: UserRepository
-) : ViewModelProvider.NewInstanceFactory() {
-
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return SearchListViewModel(repository) as T
-    }
-}
